requirements.txt
requests==2.25.1
octokitpy==0.15.0
pynacl==1.5.0